+++ lib/doggo/components/tab_navigation.ex
@@ -10,9 +10,9 @@ defmodule Doggo.Components.TabNavigation do
     """
     Renders navigation tabs.
 
-    This component is meant for tabs that link to a different view or live action.
-    If you want to render tabs that switch between in-page content panels, use
-    `tabs/1` instead.
+    This component is meant for tabs that link to a different view or live
+    action, each with a distinct URL. If you want to render tabs that switch
+    between in-page content panels, use `tabs/1` instead.
     """
   end
 
@@ -41,7 +41,27 @@ defmodule Doggo.Components.TabNavigation do
       </:item>
     </.tab_navigation>
     ```
+
+    ### Current Item
+
+    Each item has a `value` attribute that can be either a single value or a
+    list of values. If you patch between live actions using this component, you
+    would set the value to the list of live actions for which this tab is
+    active.
+
+    The root element itself has a `current_value` attribute. To determine
+    whether a tab item is active, the current value is compared with the value
+    of the tab item. If the value is a list, the tab item is considered active
+    if the current value is contained in that list.
+
+    Tab items are marked active by setting `aria-current="page"`. You can select
+    the item in CSS with `.tab-navigation a[aria-current]`.
     """
+  end
+
+  @impl true
+  def css_path do
+    "components/_tab-navigation.scss"
   end
 
   @impl true
@@ -49,7 +69,7 @@ defmodule Doggo.Components.TabNavigation do
     [
       type: :navigation,
       since: "0.6.0",
-      maturity: :developing,
+      maturity: :refining,
       modifiers: []
     ]
   end
@@ -61,17 +81,21 @@ defmodule Doggo.Components.TabNavigation do
         default: "Tabs",
         doc: """
         Aria label for the `<nav>` element. The label is especially important if you
-        have multiple `<nav>` elements on the same page. If the page is localized,
-        the label should be translated, too. Do not include "navigation" in the
-        label, since screen readers will already announce the "navigation" role as
-        part of the label.
+        have multiple `<nav>` elements on the same page, since it allows users
+        to differentiate between different navigation sections.
+
+        If the page is localized, ensure that the label is translated. Avoid
+        using the word "navigation" in the label, since screen readers will
+        already announce the "navigation" role. Additionally, the label should
+        begin with a capital letter to ensure the correct inflection when read
+        by screen readers.
         """
 
       attr :current_value, :any,
         required: true,
         doc: """
         The current value used to compare the item values with. If you use this
-        component to patch between different view actions, this could be the
+        component to patch between different view actions, this should be the
         `@live_action` attribute.
         """